:construction: [fix] Reconfig all folders
