+++ src/controllers/questionsControllers/listQuestions.controller.ts
@@ -1,31 +1,31 @@
-import { AppResponse } from "../@types";
-import { Questions } from "../models/Questions";
+import { AppResponse } from "../../@types";
+import { Questions } from "../../models/Questions";
 import { Request, Response } from "express";
 import { getCustomRepository } from "typeorm";
-import QuestionsRepository from "../repositories/question.repository";
+import QuestionsRepository from "../../repositories/question.repository";
 
-export default class ListQuestionsDelectedController {
+export default class ListQuestionsController {
   async handle(request: Request, response: Response<AppResponse<Questions[]>>) {
     try {
       const questionRepository = getCustomRepository(QuestionsRepository),
-        getAllQuestionsDelected = await questionRepository
+        getAllQuestions = await questionRepository
           .createQueryBuilder()
           .select("questions")
           .from(Questions, "questions")
           .where({
-            status: "delected",
+            status: "Active",
           })
           .getMany();
-      if (getAllQuestionsDelected.length >= 0) {
+      if (getAllQuestions.length >= 0) {
         return response.status(200).json({
           success: true,
           message: "All Questions",
-          data: getAllQuestionsDelected,
+          data: getAllQuestions,
         });
       }
       return response
         .status(400)
-        .json({ success: false, message: "No Question Delected!" });
+        .json({ success: false, message: "No Question Registred!" });
     } catch (err) {
       return response
         .status(500)
