+++ src/controllers/questionsControllers/questionsDelected.controller.ts
@@ -1,31 +1,31 @@
-import { AppResponse } from "../@types";
-import { Questions } from "../models/Questions";
+import { AppResponse } from "../../@types";
+import { Questions } from "../../models/Questions";
 import { Request, Response } from "express";
 import { getCustomRepository } from "typeorm";
-import QuestionsRepository from "../repositories/question.repository";
+import QuestionsRepository from "../../repositories/question.repository";
 
-export default class ListQuestionsController {
+export default class ListQuestionsDelectedController {
   async handle(request: Request, response: Response<AppResponse<Questions[]>>) {
     try {
       const questionRepository = getCustomRepository(QuestionsRepository),
-        getAllQuestions = await questionRepository
+        getAllQuestionsDelected = await questionRepository
           .createQueryBuilder()
           .select("questions")
           .from(Questions, "questions")
           .where({
-            status: "Active",
+            status: "delected",
           })
           .getMany();
-      if (getAllQuestions.length >= 0) {
+      if (getAllQuestionsDelected.length >= 0) {
         return response.status(200).json({
           success: true,
           message: "All Questions",
-          data: getAllQuestions,
+          data: getAllQuestionsDelected,
         });
       }
       return response
         .status(400)
-        .json({ success: false, message: "No Question Registred!" });
+        .json({ success: false, message: "No Question Delected!" });
     } catch (err) {
       return response
         .status(500)
